refactor: split mandate reference ingestion dtos
